[DotNetCore] Show project templates if MSBuild has .NET Core SDK files
If the .NET Core SDK is installed then the .NET Core project templates
are displayed in the New Project dialog.

If the .NET Core SDKs folder exists inside the MSBuild directory
and the .NET Core runtime is installed then the .NET Core project
templates are displayed.

If the .NET Core SDKs folder exists inside the MSBuild directory but
the .NET Core runtime is not installed then only show the .NET Standard
Library project templates since these do not require the runtime.