[DotNetCore] Support .NET Core SDK MSBuild files provided by Mono
Do not show the .NET Core SDK not installed dialog if Mono's MSBuild
includes the .NET Core SDKs folder. Also show a dialog indicating
the .NET Core runtime is not installed when trying to run a project
if it is not installed. Otherwise it is not obvious what the actual
problem is when running or debugging a project.