[DotNetCore] Remove F# build workaround
When FSharp.NET.Sdk 1.0.0 was used by a project the build would fail
without the MSBuildExtensionsPath being overridden to point to the
correct folder where the .NET Core sdk was installed. Creating a new
F# project uses FSharp.NET.Sdk 1.0.* which installs the latest
version which currently is 1.0.2. FSharp.NET.Sdk 1.0.1 and above
now use the MSBuildSDKsPath instead of the MSBuildExtensionsPath so
the workaround is no longer needed.