[Core] Fix new F# files not being saved in .NET Core F# project
Loading a .NET Core F# project and adding a new F# file to the project
would not add the new file to the .fsproj. The problem was that on
saving the project the default file glob for .fs files was found even
though it was excluded by the EnableDefaultCompileItems being set to
false.