[DotNetCore] Fix new F# file not added to new F# project
After creating a new F# .NET Core project, waiting for the restore
to finish, and then adding a new F# file would not add the file
to the project. This is because the main F# Sdk MSBuild imports
are added with PackageReferences in the project so when the
project is first created the default MSBuild file globs are used.
To fix this the project is re-evaluated after the NuGet packages
are restored to ensure the correct conditions are applied to the
file globs.

There is still a problem here if a new F# file is added whilst
the NuGet package restore is being run since it will not use
the updated file globs.