[DotNetCore] Fix ToolsVersion being added to project
Creating a new .NET Core project, then adding a new file to the
project was adding the ToolsVersion attribute into the saved
project file. This was due to the ToolsVersion being read again
when the project was re-evaluated and the original value from the
project, where the ToolsVersion would not be set, was being
changed to the value currently used on the MSBuildProject. So now
when re-evaluating the project the project header is not read
again by the .NET Core project extension to avoid changing the
original ToolsVersion value.