Add support for defining default policies
This patch introduces a new Default policy set, which defines the default
policies when there are no solution-defined or user-defined policies.

Before this patch, default policies were obtained by just creating an intance
of the policy type. This is still the case when a policy is not set for the
Default policy set.

This new Default policy set is mostly useful for defining default policies
for specific file types (which was impossible to do).