[Core] Fix build error when .NET Core project is referenced
When a non .NET Core project references a .NET Core project the
build would fail with an error similar to:

Error MSB4019: The imported project
"~/Library/Caches/VisualStudio/7.0/MSBuild/8209_1/Sdks/Microsoft.NET.Sdk/Sdk/Sdk.props" was not found.

The MSBuildSDKsPath was not defined when building the project that
references the .NET Core project. Now a check is made of the
referenced projects and the MSBuildSDKsPath is set there is a
.NET Core project referenced.