Fixed 'Bug 53380 - [Webtools] Editor inserts BOMs sometimes'
In old versions we used a bool to determine if a BOM could be written.
The new editor structure just uses the encoding which can control if a
bom is emitted or not. The text file utility for auto detection now
correctly gives back the right encoding. The bool for the BOM was
removed - it's no longer in the text interfaces as well. Doesn't make
sense to have that on the low level API.