[Core] Improve load times for ASP.NET Core projects
Fixed bug #54342 - Long load times for .NET Core projects with
node_modules
https://bugzilla.xamarin.com/show_bug.cgi?id=54342

If an ASP.NET Core project uses say Gulp that will add lots of files
into the node_modules directory. This was causing a long load time.
The slow part was the DefaultMSBuildEngine's Evaluate method which
with 4200+ files in the node_modules directory would take around 2.5s
for each Evaluate method call. On loading a solution with one project
the Evaluate method would be called 6 times. Without the node_modules
directory the Evaluate method would take about 150ms for each call.

To fix this a regex is created when adding items for the excluded
directories. Directories and files inside these excluded directories
are ignored. For .NET Core there are wildcard includes and removes
which are now prevented from looking for files in the node_modules
directory. The fix is not perfect - the remove items happen to have
the same wildcard as the included items so the correct directory
exclude regex can be found with an exact match when looking at the
globs for the project. With this fix the Evaluate method now takes
the same time (about 150ms) as if there is no node_modules directory.