Use the LargeAddressAware NuGet package to reliably find editbin.exe. (#2098)
The problem is that editbin.exe is only found inside VS 2017 install path and that path isn't even well defined (unlike in previous VS releases). Additionally if we want to build on CI servers that path may not be available.

After much deliberation on where to get editbin.exe I've decided to create a NuGet package "LargeAddressAware" that downloads editbin.exe (and link.exe which is required, editbin.exe just calls into link.exe) and has a targets file that injects a call to editbin after CoreCompile is done.

The target will only run on Windows (OS != Unix) and only when the LargeAddressAware MSBuild property is set.