Fix serialization of user policies
Fixed bug when getting the Default policy. It was looked up using the
policy name instead of the Id. As a result, default policies were not
properly loaded.

User policies are now serialized using the system default policy set as diff
base, so only values that are different WRT the default set are serialized.
If the values of the default set change, then users will get the new values,
unless they were explicitly changed.

Use the Mono text formatting style as default style.