[DotNetCore] Support refreshing the solution window
If a file is copied into the project folder, or Save As is used with
a file open in the editor, the file will not appear in the Solution
window until the project is reloaded. Currently there is no file
watcher support so for now the Refresh menu when you right click
a project in the Solution window will re-evaluate the project so
the files added will be displayed.