[NuGet] Fix .NET Core project build error after location changed
Fixed bug #54327 - .NET Core project build fails when solution's
location is moved
https://bugzilla.xamarin.com/show_bug.cgi?id=54327

Creating a new ASP.NET Core project, building it, closing the
solution, moving the solution to another folder, then opening it and
building it would show lots of build errors. The problem is that
the generated projectname.nuget.g.props file specifies the full path
to the project.assets.json file. The project.assets.json file
defines the NuGet packages and assemblies referenced. Since the
path to this file is now incorrect on moving the solution no
references are found by MSBuild since it cannot find the
project.assets.json file. Visual Studio handles this by always
restoring the solution when it is opened. So we now do the same for
.NET Core projects.