[Ide] Fix type system references not updated after NuGet restore
On opening a solution containing a .NET Core project that has had its
parent folder changed outside the IDE the NuGet restore would fix
the reference information but the text editor would still show
errors about missing types. This was because the project references
changed event occurred whilst the type system was loading the
project information due to a new solution being opened. The project
was added to the modified projects list but then this list is cleared
and the project is not reloaded. This leaves the type system
information out of date since no references are resolved by the
project builder initially until the NuGet restore refreshes it.