Fix bug #54691 - Preferences aren't stored
Sanitized a bit the policy container code. There Get(Type) and Get<T> methods
were supposed to have the same behavior, but they had different
implementations. Now all Get methods are based on the same code.

The real fix is that inheritsSet and inheritsScope are serialized only when
a policy was really found for that set and scope. Until now it was always
stored, so when deserializing, if the policy wasn't in the set for that
scope, it would crash.