[AspNet] Fix MVC project templates after default policy changes
Fixed bug #55012 - New ASP.NET projects do not compile
https://bugzilla.xamarin.com/show_bug.cgi?id=55012

The default naming policy was changed for new files so it matches
Visual Studio: f8d5ae2502c517c3526a954ab65dff0eddf55e9b

This change broke the ASP.NET MVC project templates since they expect
the files to have specific namespaces and now no longer compile.
The ASP.NET Project templates now use the RootNamespace instead of
the Namespace property in certain file templates so the expected
namespaces are created. The RootNamespace is the default namespace
used by the project if the file is created in the project's directory
and not in a sub-folder. This also means that the project templates
should now work for any namespace policy configured.