[Core] use 64bit if 32bit is not preferred/required
This seems like a saner way of approaching the problem of
'should Xamarin Studio start 64bit or 32bit'. The PE header
has a clear set of values we can base our decision on.
* Is the assembly compiled for x86? Use 32bit
* Is the assembly compiled for x64? Use 64bit
* Is the assembly platform agnostic and prefers 32bit? Use 32bit
* Is the assembly platform agnostic and *does not* prefer 32bit? Use 64bit

This change implements the above logic, which will allow us to easily
choose the correct bitness of the process Xamarin Studio starts by
setting the correct set of flags on our binary.