Fixed 'Bug 53380 - [Webtools] Editor inserts BOMs sometimes'
It's a fairly complex bug to fix. The behavior of VS.NET and XS
clearly differs. VS just uses the system default locale to dertmine
encoding. That's Encoding.Default on mono that's always UTF8 with bom.
So the vs loading routines assume that ascii files are always UTF8+BOM
on mono. Unfortunately we need to be able to consume the locales and
sometimes it clashes with UTF8 - the text file utility contains logic
to prevent these clashes. CP1252 files would be converted to UTF8
otherwise - it's not just a BOM bug - it's more an encoding
detection+incompatibilty bug - mosly caused by a bad design decision
from VS.NET.

Only way to fix that is using the text file utililty logic - it has
work arounded that issue since years.