Fix race in project builder
Don't reuse the workDoneEvent event anymore. In some cases it may
happen that the event is signaled twice (once when the work is
finished, and maybe once if the work is aborted just after it
finished. In those cases the second time the event is signaled it
could actually signal that the next task has been completed, which is
wrong. Also improve handling of thread abort exceptions, those
exceptions should never generate an error.

Should fix bug #52645 - XS cannot recover if the external MSBuild
builder dies