[Core] Memoize ConditionFactorExpression
When a ConditionRelationalExpression is evaluated, we have a best case scenario of a token evaluated 2 times, with a worst case of a token evaluated 3 times. Memoizing the value of the string evaluation allows us to only evaluate an item once.
Reduces total time spent in SafeParseAndEvaluate from 24565ms to 19523ms.