[Ide] New project dialog not displayed when modified file saved
Fixed bug #55457 - Issue when trying to create a new project while a
file has changes
https://bugzilla.xamarin.com/show_bug.cgi?id=55457

With a file modified and open in the text editor, selecting New
Solution shows a dialog asking to save the modified file. Clicking the
Save button would then not open the New Project dialog. The problem
was that the Document.Save method returns a task and the code
was checking the Document's IsDirty property before the save
completed which prevented the New Project dialog from being opened.