[Svn] Make unix svn resilient to native crashes
Formerly, this code would use library paths to open libsvn and libapr accordingly.

This meant that it would use the search paths for each library that would end up loading.

In the case of Xcode not being located and svn being installed via brew, it would load mismatched dlls for the two
libraries.

dlopen-ing libsvn will ensure the right apr is loaded for the probed svn lib

Bug 47563 - Visual Studio for mac crashes on opening a solution