[NuGet] Fix updating NuGet package not updating project.assets.json
Fixed bug #55563 - Updating NuGet package in .NET Core project does
not update project.assets.json file
https://bugzilla.xamarin.com/show_bug.cgi?id=55563

With an old version of a NuGet package installed in a .NET Core
project, updating the NuGet package would update the version correctly
in the project file but the project.assets.json file would have the
old version. This resulted in the wrong version displayed in the
Solution window. The problem was that the old package reference
version could be picked up since all package references were taken
from the evaluated items, as well as those defined in the project,
and only the first one is used when updating the project.assets.json
file. So now only imported package references are taken from the
evaluated items which ensures the correct package reference versions
are used from the project file.