[Core] Remove ConditionFactor memoization.
Long story short, having a memoization cache on
ConditionFactorExpression level would mean that we keep the
IExpressionContext alive (in this case an MSBuildEvaluationContext).

The problem is that these conditions are already cached at MSBuildEngine
level. Thus, we keep too much in memory, increasing XS memory over
subsequent builds by a lot.

Having some sort of signal from IExpressionContext that it is done with
and can be removed from the cache can help bring back the previous CPU
optimization.

Bug 55625 - Memleak with TensorflowSharp