[Core] Fix setting C# build action to none ignored for .NET Core
Fixed bug #55560 - Setting Build Action to None for C# file is ignored
https://bugzilla.xamarin.com/show_bug.cgi?id=55560

In a .NET Core project changing the build action for a C# file to
None made no changes to the project file. On re-opening the project
the file would have the build action of Compile.

The .NET Core sdk defines globs similar to the following:

<None Include="**/*" />
<None Remove="**/*.cs" />
<Compile Include="**/*.cs" />

On changing a C# file's build action to None the project model was
finding the file is included in the None glob include and not
making any changes to the project file. However C# files are removed
from the None glob include by the None remove glob and this was not
being taken into account. Now the remove glob is associated with its
corresponding glob include. The DefaultMSBuildEngine's
FindGlobItemsIncludingFile now does not return the None glob include
for the C# file with its build action set to None. Which results in
the project file being correctly updated.
since it is excluded by the remove.