[Ide] Fix removing project/assembly references in References dialog
Fixed bug #55747 - Removing Reference in the Edit Reference Dialog
does not work
https://bugzilla.xamarin.com/show_bug.cgi?id=55747

Removing project or assembly references from the All tab in the
Edit References dialog was not updating the project file. The
references would remain in the project. The problem was that no
matches were found for the reference information being supplied by
the All tab to the reference dialog.

The project name is now used when removing a project reference.
Previously the project's filename was used. The assembly filename
without the path and without the file extension is now used when
removing an assembly reference. Previously the assembly's full
filename was used.