[Core] Handle null dictionary value in BinaryMessage
Fixed bug #55751 - [iOS] Uploading iOS Forms PCL app from Xamarin
Studio to Test Cloud fails build step with "Unknown MSBuild Failure"
https://bugzilla.xamarin.com/show_bug.cgi?id=55751

Sending a binary message that contained a Dictionary<string, string>
and one of the values was null would result in an exception:

System.ArgumentNullException: Value cannot be null.
Parameter name: value
  at System.IO.BinaryWriter.Write (System.String value) in mcs/class/referencesource/mscorlib/system/io/binarywriter.cs:420
  at MonoDevelop.Core.Execution.BinaryMessage.WriteArray (System.IO.BinaryWriter bw, System.Object val) in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Core.Execution/BinaryMessage.cs:260

The dictionary is converted to two string arrays. If one of the
strings is null then the serialization would fail. To handle this
the null string is converted to an empty string on writing to the
stream.