[C#] Optimize completion data insertion
This change refactors CompletionResult a bit so it exposes the real list as a List<T>. This will help on introducing the CompletionResult data on completion trigger by only allocating the real list once (AddRange ICollection<T> optimization).
On the other hand, use an internal dictionary to make insertion an O(1) operation rather than O(n).