[DotNetCore] Do not show .NET Core category if runtime not installed
With Mono's MSBuild having the .NET Core SDK files, but the .NET
Core runtime or SDK not being installed separately, the .NET Core
category was displayed and it contained the .NET Standard library
project template. Whilst this project can be compiled it is
confusing that it is displayed in the .NET Core category when the
.NET Core sdk or runtime has not been installed.

The .NET Standard Library project template is now no longer displayed
in the .NET Core category if .NET Core is not installed separately.
This then prevents the .NET Core category from being displayed.
The .NET Standard Library project is still available in the
Multiplatform - Library category.