[Ide] Use the correct TreeModel in our data funcs
We should use the Model we're provided, not an arbitrary
model from somewhere else. They will not always be the
same thing!

Fixes a null ref on this line as the GetValue call
returned null because the iter was part of another
model:

`bool didRead = (bool)model.GetValue (iter, DidReadColumn);`