[Ide] Set `store` before changing the model
If someone accidentally uses this field instead one
of the gtk data funcs at least it will contain the
correct object now.

Before this change the treeview would update as soon
as the new model was set, which meant there was a time
where `treeviewSearchResults.Model` and `store` were
different objects.