[DotNetCore] Add .NET Core 2.0 project templates
These are only visible if .NET Core 2.0 sdk is installed.

Currently NuGet restore will fail for these project templates. The
NuGet.Config file is created in the project directory so it is ignored
and the needed package source is not used. Also the project itself
does not have a RuntimeFrameworkVersion so the restore fails even
with the package source enabled.