[Ide] Move template's NuGet.Config to solution directory
This is a temporary workaround for the .NET Core 2.0 project templates
that include a NuGet.Config file. This is by default created in the
project directory so it is ignored by NuGet when restoring packages.
So the NuGet.Config file is now moved to the solution directory if
it does not already exist.