[NuGet] Allow imported package references to be modified
Allow a package reference that is imported to be modified before
it is returned. This is a hack that can be used by the .NET Core
addin to change the Microsoft.NETCore.App package reference from the
stable version 2.0, which will not restore, to the actual pre-release
version that is installed. This version will then be added to the
project.assets.json file instead of the stable version.