[DotNetCore] Allow .NET Core 2.0 pre-release projects to restore
If the latest version of the .NET Core runtime installed is a 2.0
pre-release then the default Microsoft.NETCore.App version, which
will be 2.0, is modified before it is written to the
project.assets.json file so it is the pre-release version. This allows
the NuGet restore to work. This workaround should be removed when
.NET Core 2.0 stable is released.