[DotNetCore] Do not check for Sdk.props file
Fixed bug #56091 - Misleading errors about missing Sdk.props file
https://bugzilla.xamarin.com/show_bug.cgi?id=56091

When checking if the Sdk for a project exists in the .NET Core SDK
install location just check for the directory. This makes the check
consistent with the project model and also allows an sdk to not have
an Sdk.props or Sdk.targets file. Also removed logging about missing
Sdk.props files which now that the .NET Core sdk ships with Mono is
misleading.