[DotNetCore] Fix unable to run or debug ASP.NET Core project
Fixed bug #56107 - Cannot run/debug ASP.NET Web project when empty
.NET Core sdk directory
https://bugzilla.xamarin.com/show_bug.cgi?id=56107

To repro:

1) Install .NET Core SDK 1.0.3
2) Create an empty directory 1.1.1-preview2-003177 inside /usr/local/share/dotnet/sdk
3) Open VS for Mac.
4) Create an ASP.NET Core web app.

The problem was that if the Sdk imports were not found in the .NET
Core SDK then the OutputType was not read from the evaluated MSBuild
properties even though they were being set by the sdk imports that
are included with Mono. This bug could also occur if only the .NET
Core runtime is installed.