[Ide] Support different template id and group ids
When using the Microsoft templating engine, used for .NET Core
projects, it is now possible to explicitly specify the template
id and use a different group id. This allows the same project
template to be used in the same category but with a different name
and different grouping. This will be used to add a .NET Core class
library project template next to a .NET Standard project template
in the same category. These are the same underlying project templates
but take different parameters to generate different projects.

In the project template extension there is a new templateId attribute
which can be used to override the default id when looking for the
template in the NuGet package. There is also a groupId attribute
which can override the group id taken from the NuGet package for the
project template.