[DotNetCore] Support .NET Core 2.0 in project options
For a .NET Core project the project options now lists the installed
runtimes available in the target framework list. For .NET Standard
projects if the 1.x runtimes are installed then the .NET Standard
1.0 through to 1.6 is shown as available in the target framework
list. If the .NET Core 2.0 runtime is installed then .NET Standard
2.0 will be shown in the target framework list.

If the project's target framework is not an installed runtime then
in the project options the target framework will indicate that
it is not installed.