[Ide] Fix TypeInitializationException if no templates available
The new templating engine used for .NET Core projects throws an
exception if an attempt is made to save the templating cache when
there are no templates available. This then causes the MonoDevelop.Ide
addin to not load due a TypeInitializationException being thrown
when creating the MicrosoftTemplateEngineProjectTemplatingProvider.

System.Reflection.TargetInvocationException: Exception has been thrown by the target of an invocation.
  ---> System.TypeInitializationException: The type initializer for
'MonoDevelop.Ide.Templates.MicrosoftTemplateEngineProjectTemplatingProvider' threw an exception.
  ---> System.NullReferenceException: Object reference not set to an instance of an object
  at Microsoft.TemplateEngine.Edge.Settings.SettingsLoader.Save (Microsoft.TemplateEngine.Edge.Settings.TemplateCache cacheToSave) in
  at Microsoft.TemplateEngine.Edge.Settings.SettingsLoader.Save ()
  at MonoDevelop.Ide.Templates.MicrosoftTemplateEngineProjectTemplatingProvider.UpdateCache () in
  main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Templates/MicrosoftTemplateEngineProjectTemplatingProvider.cs:85
  at MonoDevelop.Ide.Templates.MicrosoftTemplateEngineProjectTemplatingProvider..cctor () in
  main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Templates/MicrosoftTemplateEngineProjectTemplatingProvider.cs:66

Now if there are no templates then no attempt is made to update the
templating cache which avoids this exception and prevents extra
work being done when it is not needed. If a template becomes available
the cache will be updated.