[DotNetCore] Fix wrong preview NETStandard.Library version used
When restoring the NuGet packages for a .NET Standard 2.0 project
the NETStandard.Library version being used was 1.6.1 which is from
the SDK files that ship with Mono's MSBuild. As a workaround the
Microsoft.NETCoreSdk.BundledVersions.props file is read and the
BundledNETStandardPackageVersion property is used to fix the
NETStandard.Library package reference when the project.assets.json
file is created. This package reference change is only done if the
project targets netstandard2.0.