[Ide] Support template conditions in new templating engine
A condition can be assigned to a template which determines whether it
is used to create the project. The conditions are tested against the
parameters defined, usually by a template wizard, and the correct
template is used. This allows one template to be grouped with others
and depending on the parameters defined it can be selected.