[Ide] Handle grouped templates with same template id
The .NET Core 2.0 and 1.0 project templates are grouped but have the
same template id. Using the same template id means that duplicate
templates are not listed in the recently used list in the New
Project dialog. However this would sometimes result in the New
Project dialog not being able to find a template that supported a
condition if the recently used template was selected. This would
happen if an F# ASP.NET Core Web App project template was selected
and .NET Core 2.0 was chosen in the project template wizard. To
handle this when a template cannot be found using the currently
selected template the New Project dialog falls back to checking
the templates that have the same id and category and also support
the current condition.