[DotNetCore] Improve display names for target frameworks.
Target framework names displayed in the project options and the
project template wizard have been modified to be consistent with
Visual Studio on Windows.

".NETCoreApp 1.1" is now displayed as ".NET Core 1.1"
".NETStandard 1.1" is now displayed as ".NET Standard 1.1"

Currently these names are done in code. In the future it may be
better to implement the /MonoDevelop/Core/Frameworks extension
point for .NET Core and .NET Standard target frameworks where a
user friendly name can be specified in an .xml file.