Added MonoDevelop SDK resolver
Created a new SDK resolver that is used to find SDKs defined by add-ins.
This resolver needs to be copied to MD's local msbuild environment,
so that projects can be properly built.

Removed infrastructure for assigning the SDKs path to project builders.
This is not required anymore since SDKs are now found using resolvers.

In MSBuildEvaluationContext, don't define MSBuildSDKsPath. This is
now handled while evaluating.