Fixed 'Bug 55735 - Intellisense popups are layered, intrusive, and disruptive'.
Basically that bug was caused by the async nature of updating the
parameter info window. The window size may update after the
PositionParameterInfoWindow call so it operates on invalid values.
However it's more natural to 'correct' the x/y position after the
resizing so the location correction algorithm got splitted up. Seems
to fix the issue for me.