[Core] Fix API compatibility.
That broke due to the vs data structure port. A newOffset was added
w/o changing the old constructors. All creators of the
TextChangeEventArgs should be aware of that - there is no reason that
text change event args constructors are public. They're only created
by the editor or code that has internal access.