[DotNetCore] Do not restore when project being re-evaluated
If a .NET Core project references a shared project then when the
.NET Core project is re-evaluated the shared project will trigger
reference change events. Running a restore when this occurred would
cause the restore to get stuck in a loop and be run repeatedly.
Now if the project is being re-evaluated then a restore is not run.
Re-evaluation is done during the NuGet restore process anyway so
there is no need to trigger another restore during a re-evaluation.