[Core] Fix error re-evaluating project referencing a shared project
Fixed bug #56147 - NRE restoring Shared Project in a NetCore app
https://bugzilla.xamarin.com/show_bug.cgi?id=56147

Calling Project.ReevaluateProject on a project that references a
shared assets project would throw a null reference exception.

System.NullReferenceException: Object reference not set to an instance of an object
  at MonoDevelop.Projects.SharedAssetsProjects.SharedAssetsProject.HandleReferenceRemovedFromProject (System.Object sender, MonoDevelop.Projects.ProjectReferenceEventArgs e)
  in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Projects.SharedAssetsProjects/SharedAssetsProject.cs:357
  at (wrapper delegate-invoke) <Module>:invoke_void_object_ProjectReferenceEventArgs (object,MonoDevelop.Projects.ProjectReferenceEventArgs)
  at MonoDevelop.Projects.Solution.OnReferenceRemovedFromProject (MonoDevelop.Projects.ProjectReferenceEventArgs args)
  in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Projects/Solution.cs:1175

During re-evaluation the shared file is removed from the project
before the reference change event is fired. The project file did not
exist in the project and the shared assets project was not handling
this.

Fixing the null reference was not enough since after the re-evaluation
the project did not have its original files in the Project.Files
collection. This was because the SharedAssetsProjectMSBuildExtension's
OnReadProject method calls the Project's LoadProjectItems and
removes the original project's items since it calls SetItems on the
Items collection during an re-evaluation. No files were added by
the SharedAssetsProjectMBuildExtension since it did not evaluate
the MSBuildProject for the .projitems so this code has been removed.
The files are added later on either when the project is bound to
the solution during a normal project load or when the reference
change event is fired during a re-evaluation.