[DotNetCore] Do not show files from shared project in Solution pad
Fixed bug #56144 - NetCore Projects show the files from a referenced
Shared Project
https://bugzilla.xamarin.com/show_bug.cgi?id=56144

When a .NET Core project referenced a shared assets project the files
from the shared project were being displayed in the Solution window
when they should not have been displayed. The .NET Core project
extension was adding these imported files when the project was loaded.
Now the imported files are checked to see if they belong to a shared
project, indicated by the MSBuild project having the SharedGUID and
HasSharedItems properties, and if so they are ignored by the .NET
Core project extension, but will be added as Hidden, DontPersist
files by the shared assets project. This results in the files not
being displayed in the Solution window.