[Core] Fix project options cannot be opened for generic projects
On trying to open the project options dialog for a generic project
a null reference exception was thrown by the RunConfigurationPanel
since it could not find an run configuration editor for generic
projects. To avoid this the generic project now indicates it does
not support run configurations so the project options dialog can
be opened.