[DotNetCore] Fix failure to create project templates
Fixed bug #56751 - New templates for .NET Core Support cannot be created
https://bugzilla.xamarin.com/show_bug.cgi?id=56751

With .NET Core 2.0 installed it was not possible to create the ASP.NET
Core Razor Pages project template. The problem was that since the
wizard is not displayed when no target framework can be selected the
parameter used in the template condition was not being set so the
template could not be found. This resulted in an error in the IDE log
and the New Project dialog remaining open. Now if a wizard is not
going to be displayed the parameters are set based on what .NET Core
versions are installed so the correct template can be resolved.