Fixed 'Bug 56650 - Many workspace errors logged while coding'
Without a repro it's hard to tell why that happens. This patch should
ensure that no invalid document is open and that the analysis document
invalidates when the roslyn workspace closes the open document. The
source text container has now a check on invalid documents as well and
tries to reopen the document in that case. This should fix the issue.