[Ide] Improve performance of first folder expansion.
WHenever we open a new folder, we calculate project paths to decide where the file ends up.
That said, the ProjectFileCollection is a bit heavy, as it does a lot more than just being a container for ProjectFile. It creates a tree of items, allocates a lot of data that we don't need, subscribes to change events. Addition of an item to a ProjectFileCollection is not O(1) either, its a recursive method which queries dictionaries based on how many path segments there are. The code just needs to check whether the path is a child of another path, so, since we do it in the GetFolderContent method, just use a simple List<ProjectFile>.
Expanding the icons folder from MonoDevelop.Ide is now ~instant, from 2-3s observed time.